requirements.txt
Django==1.9
django-filter==0.11.0
django-rest==0.0.1
django-webpack-loader==0.2.3
djangorestframework==3.3.1
Markdown==2.6.5
wheel==0.24.0
